Create Banking System-OOPS
The Banking System with Checking and Savings Accounts, implemented using Object-Oriented Programming (OOP) principles, is a Java-based application that models the operations of a bank. It features distinct account types, customer management, balance tracking, transaction history, and interest calculation for savings accounts, all designed with the principles of OOP in mind, providing a flexible and organized framework for managing financial transactions.